Add files via upload
24V Battery Level İndicator
24V BATTERY LEVEL



The circuit I designed is suitable for 24V batteries. You include the + and - ends of the battery in the connector. If the red led is on, your battery is low. If the yellow led is on, your battery is moderately charged. If the blue LED is on, your battery is fully charged.



I designed it for you as a very simple PCB to solder and use. Thanks.



Zener diode is composed of P and N type semiconductor materials, it is made of silicon. They are diodes that keep the voltage applied to their ends constant. Zener diodes do not pass current up to a certain voltage value. This voltage is called the break or zener voltage (Vz). When connected to the circuit in the right direction, it works like a normal diode. When it is connected in the reverse direction, it does not conduct until the breakdown voltage (zener voltage), and when the breakdown voltage is exceeded, it passes current in the form of an avalanche effect and turns on. When the reverse voltage is removed, the zener diode returns to its normal state. Since Zener diodes work in reverse polarity, they are reverse connected to the circuit. The Zener point (breakdown voltage) value is determined by adjusting the amount of additive during the production phase.